refactor: split code for more legibility
- create basics for draw simple types
- create utils for get methods
- math types has already a cool toString method
