+++ ECSInspector/Scripts/UImGuiEntityInspectorWindow.Basic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be5abef199b7b334c915858b4abf278c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
