+++ ECSInspector/Scripts/UImGuiEntityInspectorWindow.Utils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 72a94235a37523f4e9b386a14dc2c6fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
